doc, tools: extract render
